// src/types/product.ts
export interface Product {
  id: number;
  serialNumber: number;
  isNew: number;
  photo: string;
  title: string;
  type: string;
  specification: string;
  guarantee: {
      start: string;
      end: string;
  };
  price: {
      value: number;
      symbol: string;
      isDefault: number;
  }[];
  order: number;
  date: string;
}